feat: Add stat definitions and computeStatValue function for player statistics management
