+++ src/client/StatisticsModal.ts
@@ -1,9 +1,10 @@
-import { LitElement, html, svg } from "lit";
+import { LitElement, PropertyValues, html, svg } from "lit";
 import { customElement, property, query, state } from "lit/decorators.js";
 import { PlayerType, UnitType } from "../core/game/Game";
 import { GameView, PlayerView } from "../core/game/GameView";
 import { getTechNodes, type Category } from "../core/tech/ResearchTree";
 import "./components/baseComponents/Modal";
+import { AVAILABLE_STATS, computeStatValue } from "./stats/StatDefinitions";
 import statsStore from "./stats/StatsStore";
 
 @customElement("statistics-modal")
@@ -42,10 +43,7 @@ export class StatisticsModal extends LitElement {
       clearInterval(this._intervalId);
       this._intervalId = null;
     }
-    if (this._graphActiveMetric) {
-      statsStore.stop(this._graphActiveMetric);
-      this._graphActiveMetric = null;
-    }
+    // Do NOT stop sampling metrics here; we want them to continue in background
     this._stopGraphRenderLoop();
   }
 
@@ -506,22 +504,9 @@ export class StatisticsModal extends LitElement {
 
   private _ensureGraphSampling() {
     this._ensureGraphDefaults();
-    const metric = this._graphMetric!;
-    if (this._graphActiveMetric && this._graphActiveMetric !== metric) {
-      statsStore.stop(this._graphActiveMetric);
-      this._graphActiveMetric = null;
-    }
-    // Sample all players so we have data even if selection changes
-    statsStore.start(
-      metric,
-      () => this._playersAll(),
-      (m, p) => this._computeStatValue(m, p).sortValue,
-      (p) => p.isAlive(),
-      () => this.game?.ticks() ?? 0,
-    );
-    this._graphActiveMetric = metric;
+    // Sampling is now initialized in ClientGameRunner.ts
+    this._graphActiveMetric = this._graphMetric;
   }
-
   private _startGraphRenderLoop() {
     if (this._graphRenderIntervalId) return;
     const tick = () => {
@@ -529,7 +514,7 @@ export class StatisticsModal extends LitElement {
       this._ensureGraphSampling();
       this._graphRenderTick++;
     };
-    this._graphRenderIntervalId = setInterval(tick, 1000);
+    this._graphRenderIntervalId = setInterval(tick, 10000);
     tick();
   }
 
@@ -1216,6 +1201,13 @@ export class StatisticsModal extends LitElement {
     this.addEventListener("modal-close", () => this._stopAutoRefresh());
   }
 
+  protected updated(changedProperties: PropertyValues) {
+    super.updated(changedProperties);
+    if (changedProperties.has("game") && this.game) {
+      this._ensureGraphSampling();
+    }
+  }
+
   private _refreshList() {
     this._listRefreshTick++;
   }
@@ -1242,45 +1234,7 @@ export class StatisticsModal extends LitElement {
   }
 
   private _availableListStats(): string[] {
-    return [
-      "Gold",
-      "Industrial Production",
-      "Population",
-      "Workers",
-      "Troops",
-      "Productivity %",
-      "Productivity Growth %",
-      "Investment – Production %",
-      "Investment – Production Amount/s",
-      "Investment – Roads %",
-      "Investment – Roads Amount/s",
-      "Investment – Research %",
-      "Investment – Research Amount/s",
-      "Road Quality %",
-      "Road Completion %",
-      // Structures (match PlayerInfoOverlay ordering)
-      "City",
-      "Hospital",
-      "Academy",
-      "Research Lab",
-      "Factory",
-      "Port",
-      "Warship",
-      "Missile Silo",
-      "SAM Launcher",
-      "Air Field",
-      "Fighter Jet",
-      "Defense Post",
-      // Tech overview
-      "Researched Techs",
-      "Research Level",
-      // Tech by category (researched/total; sort by researched)
-      "Land Techs",
-      "Sea Techs",
-      "Air Techs",
-      "Nuclear Techs",
-      "Economy Techs",
-    ];
+    return AVAILABLE_STATS;
   }
 
   private _computeStatValue(
@@ -1292,159 +1246,7 @@ export class StatisticsModal extends LitElement {
     displayPrimary: string;
     displaySecondary?: string;
   } {
-    const gross = this.game?.config().grossGoldAdditionRate(p) ?? 0;
-    const perSecond = 10;
-    const inv = p.investmentRate?.() ?? (p as any).data?.investmentRate ?? 0;
-    const roadRate =
-      (p as any).roadInvestmentRate?.() ??
-      p.roadInvestmentRate?.() ??
-      (p as any).data?.roadInvestmentRate ??
-      0;
-    const researchRate =
-      (p as any).researchInvestmentRate?.() ??
-      p.researchInvestmentRate?.() ??
-      (p as any).data?.researchInvestmentRate ??
-      0;
-    const prodAmt = gross * inv * perSecond;
-    const roadAmt = gross * roadRate * perSecond;
-    const researchAmt = gross * researchRate * perSecond;
-    const ip =
-      (p as any).industrialProduction?.() ??
-      (p as any).industrialProduction ??
-      (p as any).data?.industrialProduction ??
-      0;
-    switch (label) {
-      case "Gold":
-        return {
-          sortValue: Number(p.gold?.() ?? 0),
-          displayPrimary: String(p.gold?.() ?? 0),
-        };
-      case "Industrial Production":
-        return { sortValue: Number(ip ?? 0), displayPrimary: String(ip ?? 0) };
-      case "Population":
-        return {
-          sortValue: p.population(),
-          displayPrimary: String(p.population()),
-        };
-      case "Workers":
-        return { sortValue: p.workers(), displayPrimary: String(p.workers()) };
-      case "Troops":
-        return { sortValue: p.troops(), displayPrimary: String(p.troops()) };
-      case "Productivity %": {
-        const val = (p.productivity?.() ?? 0) * 100;
-        return { sortValue: val, displayPrimary: `${val.toFixed(1)}%` };
-      }
-      case "Productivity Growth %": {
-        const val = (p.productivityGrowthPerMinute?.() ?? 0) * 100;
-        return { sortValue: val, displayPrimary: `${val.toFixed(1)}%` };
-      }
-      case "Investment – Production %": {
-        const val = (inv ?? 0) * 100;
-        return { sortValue: val, displayPrimary: `${val.toFixed(0)}%` };
-      }
-      case "Investment – Production Amount/s":
-        return { sortValue: prodAmt, displayPrimary: prodAmt.toFixed(2) };
-      case "Investment – Roads %": {
-        const val = (roadRate ?? 0) * 100;
-        return { sortValue: val, displayPrimary: `${val.toFixed(0)}%` };
-      }
-      case "Investment – Roads Amount/s":
-        return { sortValue: roadAmt, displayPrimary: roadAmt.toFixed(2) };
-      case "Investment – Research %": {
-        const val = (researchRate ?? 0) * 100;
-        return { sortValue: val, displayPrimary: `${val.toFixed(0)}%` };
-      }
-      case "Investment – Research Amount/s":
-        return {
-          sortValue: researchAmt,
-          displayPrimary: researchAmt.toFixed(2),
-        };
-      case "Road Quality %": {
-        const val = (p.roadNetworkQuality?.() ??
-          (p as any).data?.roadNetworkQuality ??
-          100) as number;
-        return { sortValue: val, displayPrimary: `${Math.round(val)}%` };
-      }
-      case "Road Completion %": {
-        const val = (p.roadNetworkCompletion?.() ??
-          (p as any).data?.roadNetworkCompletion ??
-          100) as number;
-        return { sortValue: val, displayPrimary: `${Math.round(val)}%` };
-      }
-      // Structures (upgradeOwned use unitsOwned, others use units().length)
-      case "City":
-      case "Hospital":
-      case "Academy":
-      case "Research Lab":
-      case "Factory":
-      case "Port": {
-        const map: Record<string, UnitType> = {
-          City: UnitType.City,
-          Hospital: UnitType.Hospital,
-          Academy: UnitType.Academy,
-          "Research Lab": UnitType.ResearchLab,
-          Factory: UnitType.Factory,
-          Port: UnitType.Port,
-        };
-        const t = map[label];
-        const count = p.unitsOwned(t);
-        return { sortValue: count, displayPrimary: String(count) };
-      }
-      case "Warship":
-      case "Missile Silo":
-      case "SAM Launcher":
-      case "Air Field":
-      case "Fighter Jet":
-      case "Defense Post": {
-        const map: Record<string, UnitType> = {
-          Warship: UnitType.Warship,
-          "Missile Silo": UnitType.MissileSilo,
-          "SAM Launcher": UnitType.SAMLauncher,
-          "Air Field": UnitType.Airfield,
-          "Fighter Jet": UnitType.FighterJet,
-          "Defense Post": UnitType.DefensePost,
-        };
-        const t = map[label];
-        const count = p.units(t).length;
-        return { sortValue: count, displayPrimary: String(count) };
-      }
-      // Tech overview
-      case "Researched Techs": {
-        const n = (p as any).data?.researchTreeTechs?.length ?? 0;
-        return { sortValue: n, displayPrimary: String(n) };
-      }
-      case "Research Level": {
-        const lvl = Number(p.researchTechLevel()) || 0;
-        return { sortValue: lvl, displayPrimary: String(lvl) };
-      }
-
-      // Tech by category (researched/total; sort by researched)
-      case "Land Techs":
-      case "Sea Techs":
-      case "Air Techs":
-      case "Nuclear Techs":
-      case "Economy Techs": {
-        const labelToCat: Record<string, Category> = {
-          "Land Techs": "Land",
-          "Sea Techs": "Sea",
-          "Air Techs": "Air",
-          "Nuclear Techs": "Nuclear",
-          "Economy Techs": "Economy",
-        };
-        const cat = labelToCat[label];
-        const nodes = getTechNodes();
-        const total = nodes.filter((n) => n.category === cat).length;
-        let researched = 0;
-        for (const n of nodes) {
-          if (n.category === cat && p.hasResearchedTech(n.id)) researched++;
-        }
-        return {
-          sortValue: researched,
-          displayPrimary: `${researched}/${total}`,
-        };
-      }
-    }
-    return { sortValue: 0, displayPrimary: "—" };
+    return computeStatValue(this.game, label, p);
   }
 }
